+++ CHANGELOG.md
@@ -10,6 +10,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Changed
 - Changes stderr to behave the same way as stdout
 
+### Removed
+- Removes logger
+
 [0.1.1](https://github.com/jshwi/spall/releases/tag/v0.1.1) - 2022-01-09
 ------------------------------------------------------------------------
 ### Fixed